chore: Reorganize Help page with modular components.
