+++ src/app/(sidebar)/help/_components/SearchBar.tsx
@@ -1,0 +1,25 @@
+import SvgIcon from "@/components/images/SvgIcon";
+
+interface SearchBarProps {
+    searchQuery: string;
+    setSearchQuery: (query: string) => void;
+}
+
+const SearchBar = ({ searchQuery, setSearchQuery }: SearchBarProps) => {
+    return (
+        <div className="relative mb-6">
+            <div className="absolute inset-y-0 left-0 pl-3 flex items-center pointer-events-none">
+                <SvgIcon name="search" size={18} useSystemTheme={true} />
+            </div>
+            <input
+                type="text"
+                className="block w-full pl-10 pr-3 py-3 border rounded-lg bg-[var(--bg-card)] border-[var(--border-color)] text-[var(--text-primary)] focus:outline-none focus:ring-2 focus:ring-blue-500"
+                placeholder="Search for help topics..."
+                value={searchQuery}
+                onChange={(e) => setSearchQuery(e.target.value)}
+            />
+        </div>
+    );
+};
+
+export default SearchBar;
